requirements.txt
openai>=1.109.0
flask>=3.1.0
flask-cors>=6.0.1
python-dotenv>=1.1.1
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Pillow==10.1.0
Werkzeug==3.1.3
psycopg2-binary>=2.9.11
vercel-blob>=0.4.2
